Rename CEF implementations to match the process they are called in
