+++ desktop/src/cef/internal.rs
@@ -1,10 +1,10 @@
-mod app;
+mod browser_process_app;
+mod browser_process_client;
 mod browser_process_handler;
-mod client;
-mod non_browser_app;
 mod render_handler;
+mod render_process_app;
 
-pub(crate) use app::AppImpl;
-pub(crate) use client::ClientImpl;
-pub(crate) use non_browser_app::NonBrowserAppImpl;
+pub(crate) use browser_process_app::AppImpl;
+pub(crate) use browser_process_client::ClientImpl;
 pub(crate) use render_handler::RenderHandlerImpl;
+pub(crate) use render_process_app::NonBrowserAppImpl;
